requirements.txt
streamlit>=1.34
pandas>=2.2
camelot-py[cv]==0.11.0     # PDF table extractor (needs Ghostscript)
tabula-py>=2.9             # fallback extractor
openpyxl>=3.1              # Excel export
matplotlib>=3.9            # bar chart (Streamlit wrapper)
fpdf2>=2.7.8               # PDF generation for synthetic data (Unicode support)
JPype1>=1.5.2              # Java bridge for tabula-py
watchdog                   # For better Streamlit file watching

